chore(EMS-1342): clean up custom keystone GQL resolvers
