+++ src/api/custom-resolvers/mutations/index.ts
@@ -1,0 +1,23 @@
+import createAccount from './create-account';
+import verifyAccountEmailAddress from './verify-account-email-address';
+import sendEmailConfirmEmailAddress from './send-email-confirm-email-address';
+import accountSignIn from './account-sign-in';
+import accountSignInSendNewCode from './account-sign-in-new-code';
+import verifyAccountSignInCode from './verify-account-sign-in-code';
+import addAndGetOTP from './add-and-get-OTP';
+import deleteApplicationByReferenceNumber from './delete-application-by-refrence-number';
+import updateExporterCompanyAndCompanyAddress from './update-exporter-company-and-company-address';
+import submitApplication from './submit-application';
+
+export {
+  createAccount,
+  verifyAccountEmailAddress,
+  sendEmailConfirmEmailAddress,
+  accountSignIn,
+  accountSignInSendNewCode,
+  verifyAccountSignInCode,
+  addAndGetOTP,
+  deleteApplicationByReferenceNumber,
+  updateExporterCompanyAndCompanyAddress,
+  submitApplication,
+};
