+++ src/api/custom-resolvers/mutations/verify-account-email-address.ts
@@ -1,8 +1,8 @@
 import { Context } from '.keystone/types'; // eslint-disable-line
 import { isBefore } from 'date-fns';
-import { FIELD_IDS } from '../constants';
-import getAccountByField from '../helpers/get-account-by-field';
-import { VerifyEmailAddressVariables, VerifyEmailAddressResponse } from '../types';
+import { FIELD_IDS } from '../../constants';
+import getAccountByField from '../../helpers/get-account-by-field';
+import { VerifyEmailAddressVariables, VerifyEmailAddressResponse } from '../../types';
 
 /**
  * verifyAccountEmailAddress
