+++ src/api/custom-resolvers/mutations/verify-account-sign-in-code.ts
@@ -1,10 +1,10 @@
 import { Context } from '.keystone/types'; // eslint-disable-line
 import { isAfter } from 'date-fns';
-import { ACCOUNT } from '../constants';
-import getExporterById from '../helpers/get-exporter-by-id';
-import isValidOTP from '../helpers/is-valid-otp';
-import create from '../helpers/create-jwt';
-import { VerifyAccountSignInCodeVariables, VerifyAccountSignInCodeResponse } from '../types';
+import { ACCOUNT } from '../../constants';
+import getExporterById from '../../helpers/get-exporter-by-id';
+import isValidOTP from '../../helpers/is-valid-otp';
+import create from '../../helpers/create-jwt';
+import { VerifyAccountSignInCodeVariables, VerifyAccountSignInCodeResponse } from '../../types';
 
 const {
   JWT: { SESSION_EXPIRY },
