+++ src/api/custom-resolvers/queries/get-companies-house-information.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import dotenv from 'dotenv';
-import { mapCompaniesHouseFields } from '../helpers/mapCompaniesHouseFields';
-import { GetCompaniesHouseInformationVariables } from '../types';
+import { mapCompaniesHouseFields } from '../../helpers/mapCompaniesHouseFields';
+import { GetCompaniesHouseInformationVariables } from '../../types';
 
 dotenv.config();
